requirements.txt
streamlit
opencv-python
numpy
pillow
pyyaml
pandas
torch
torchvision
ultralytics
ffmpeg-python
supervision
scikit-learn